Replace `ssz` with `ssz2`, adapt ssz_derive
