+++ Cargo.toml
@@ -17,7 +17,6 @@ members = [
 	"eth2/utils/serde_hex",
 	"eth2/utils/slot_clock",
 	"eth2/utils/ssz",
-	"eth2/utils/ssz2",
 	"eth2/utils/ssz_derive",
 	"eth2/utils/swap_or_not_shuffle",
 	"eth2/utils/tree_hash",
