+++ eth2/utils/ssz/src/decode/impls.rs
