--- eth2/utils/ssz2/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "ssz2"
-version = "0.1.0"
-authors = ["Paul Hauner <paul@paulhauner.com>"]
-edition = "2018"
-
-[dependencies]
-bytes = "0.4.9"
-ethereum-types = "0.5"
-hashing = { path = "../hashing" }
-int_to_bytes = { path = "../int_to_bytes" }
-hex = "0.3"
-yaml-rust = "0.4"
