--- eth2/utils/ssz2/fuzz/.gitignore
@@ -1,4 +1,0 @@
-
-target
-corpus
-artifacts
