--- eth2/utils/ssz2/fuzz/Cargo.toml
@@ -1,105 +1,0 @@
-
-[package]
-name = "ssz-fuzz"
-version = "0.0.1"
-authors = ["Automatically generated"]
-publish = false
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-ethereum-types = "0.5"
-
-[dependencies.ssz]
-path = ".."
-[dependencies.libfuzzer-sys]
-git = "https://github.com/rust-fuzz/libfuzzer-sys.git"
-
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
-[[bin]]
-name = "fuzz_target_bool_decode"
-path = "fuzz_targets/fuzz_target_bool_decode.rs"
-
-[[bin]]
-name = "fuzz_target_bool_encode"
-path = "fuzz_targets/fuzz_target_bool_encode.rs"
-
-[[bin]]
-name = "fuzz_target_u8_decode"
-path = "fuzz_targets/fuzz_target_u8_decode.rs"
-
-[[bin]]
-name = "fuzz_target_u8_encode"
-path = "fuzz_targets/fuzz_target_u8_encode.rs"
-
-[[bin]]
-name = "fuzz_target_u16_decode"
-path = "fuzz_targets/fuzz_target_u16_decode.rs"
-
-[[bin]]
-name = "fuzz_target_u16_encode"
-path = "fuzz_targets/fuzz_target_u16_encode.rs"
-
-[[bin]]
-name = "fuzz_target_u32_decode"
-path = "fuzz_targets/fuzz_target_u32_decode.rs"
-
-[[bin]]
-name = "fuzz_target_u32_encode"
-path = "fuzz_targets/fuzz_target_u32_encode.rs"
-
-[[bin]]
-name = "fuzz_target_u64_decode"
-path = "fuzz_targets/fuzz_target_u64_decode.rs"
-
-[[bin]]
-name = "fuzz_target_u64_encode"
-path = "fuzz_targets/fuzz_target_u64_encode.rs"
-
-[[bin]]
-name = "fuzz_target_usize_decode"
-path = "fuzz_targets/fuzz_target_usize_decode.rs"
-
-[[bin]]
-name = "fuzz_target_usize_encode"
-path = "fuzz_targets/fuzz_target_usize_encode.rs"
-
-[[bin]]
-name = "fuzz_target_hash256_decode"
-path = "fuzz_targets/fuzz_target_hash256_decode.rs"
-
-[[bin]]
-name = "fuzz_target_hash256_encode"
-path = "fuzz_targets/fuzz_target_hash256_encode.rs"
-
-[[bin]]
-name = "fuzz_target_address_decode"
-path = "fuzz_targets/fuzz_target_address_decode.rs"
-
-[[bin]]
-name = "fuzz_target_address_encode"
-path = "fuzz_targets/fuzz_target_address_encode.rs"
-
-[[bin]]
-name = "fuzz_target_vec_address_decode"
-path = "fuzz_targets/fuzz_target_vec_address_decode.rs"
-
-[[bin]]
-name = "fuzz_target_vec_bool_decode"
-path = "fuzz_targets/fuzz_target_vec_bool_decode.rs"
-
-[[bin]]
-name = "fuzz_target_vec_decode"
-path = "fuzz_targets/fuzz_target_vec_decode.rs"
-
-[[bin]]
-name = "fuzz_target_vec_encode"
-path = "fuzz_targets/fuzz_target_vec_encode.rs"
-
-[[bin]]
-name = "fuzz_target_vec_u64_decode"
-path = "fuzz_targets/fuzz_target_vec_u64_decode.rs"
